# docs/swagger.yaml
basePath: /
definitions:
  model.BookingCreatePayload:
    properties:
      customerObjectID:
        type: string
      customerid:
        type: string
      note:
        type: string
      serviceObjectID:
        type: string
      serviceid:
        type: string
      time:
        type: string
    type: object
  model.BookingUpdatePayload:
    properties:
      note:
        type: string
      serviceObjectID:
        type: string
      serviceid:
        type: string
      time:
        type: string
    type: object
  model.CompanyCreatePayload:
    properties:
      address:
        type: string
      email:
        type: string
      location:
        $ref: '#/definitions/model.Location'
        type: object
      name:
        type: string
      phone:
        type: string
    type: object
  model.CompanyUpdatePayload:
    properties:
      active:
        type: boolean
      address:
        type: string
      email:
        type: string
      location:
        $ref: '#/definitions/model.Location'
        type: object
      name:
        type: string
      phone:
        type: string
    type: object
  model.Coordinates:
    properties:
      acc:
        type: integer
      lat:
        type: number
      lng:
        type: number
    type: object
...
  model.Location:
    properties:
      coordinates:
        $ref: '#/definitions/model.Coordinates'
        type: object
      type:
        type: string
    type: object
...
  util.Response:
    properties:
      data:
        type: object
      message:
        type: string
    type: object
host: micasvn.ddns.net:9999
info:
  contact: {}
  description: Documentation of Garage Service API
  license: {}
  title: Garage Service API
  version: 1.0.0
paths:
  /admin/bookings:
    get:
      consumes:
      - application/json
      description: Return a list of bookings
      parameters:
      - description: Status state
        in: query
        name: status
        type: string
      - description: Service's ID
        in: query
        name: serviceid
        type: string
      - description: Customer's ID
        in: query
        name: customerid
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Booking list
      tags:
      - Admin - Bookings
    post:
      consumes:
      - application/json
      description: Create a booking
      parameters:
      - description: Booking Create Payload
        in: body
        name: BookingCreatePayload
        required: true
        schema:
          $ref: '#/definitions/model.BookingCreatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Booking create
      tags:
      - Admin - Bookings
  /admin/bookings/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a booking
      parameters:
      - description: Booking's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Booking delete
      tags:
      - Admin - Bookings
    get:
      consumes:
      - application/json
      description: Return details of a booking
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Booking detail
      tags:
      - Admin - Bookings
    patch:
      consumes:
      - application/json
      description: Update booking's details
      parameters:
      - description: Booking's ID
        in: path
        name: id
        required: true
        type: string
      - description: Booking Update Payload
        in: body
        name: BookingUpdatePayload
        required: true
        schema:
          $ref: '#/definitions/model.BookingUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Booking update
      tags:
      - Admin - Bookings
  /admin/companies:
    get:
      consumes:
      - application/json
      description: Return a list of companies
      parameters:
      - description: Name keyword
        in: query
        name: name
        type: string
      - description: Active state
        in: query
        name: active
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Phone number
        in: query
        name: phone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Company list
      tags:
      - Admin - Companies
    post:
      consumes:
      - application/json
      description: Create a company
      parameters:
      - description: Company Create Payload
        in: body
        name: CompanyCreatePayload
        required: true
        schema:
          $ref: '#/definitions/model.CompanyCreatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Company create
      tags:
      - Admin - Companies
  /admin/companies/{id}:
    get:
      consumes:
      - application/json
      description: Return details of a company
      parameters:
      - description: Company's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Company detail
      tags:
      - Admin - Companies
    patch:
      consumes:
      - application/json
      description: Update company's details
      parameters:
      - description: Company's ID
        in: path
        name: id
        required: true
        type: string
      - description: Active state
        in: query
        name: active
        type: string
      - description: Company Update Payload
        in: body
        name: CompanyUpdatePayload
        schema:
          $ref: '#/definitions/model.CompanyUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Company update
      tags:
      - Admin - Companies
  /admin/compnanies/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a company
      parameters:
      - description: Company's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: Admin API - Company delete
      tags:
      - Admin - Companies
...
securityDefinitions:
  BearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"